requirements.txt
streamlit>=1.30.0
requests>=2.31.0
pandas>=2.0.3
scikit-learn>=1.3.0
plotly>=5.18.0
nltk>=3.8.1
spacy>=3.7.2
transformers>=4.37.2
torch>=2.2.0
pyjwt>=2.8.0
